Movi el comportamiento de pantalla completa al banner para que funcione en todos los lugares donde se usa, ahora mismo solo se puede ver la imagen en pantalla completa en add-pet, no es posts o productos
